+++ packages/adapters/eventstore-adapters/eventstore-mysql/src/__mocks__/@reimagined/eventstore-base.ts
@@ -1,6 +1,0 @@
-export {
-  EventstoreResourceAlreadyExistError,
-  EventstoreResourceNotExistError,
-} from '@reimagined/eventstore-base'
-
-export default jest.fn()