+++ packages/runtime/adapters/eventstore-adapters/eventstore-mysql/src/__mocks__/@reimagined/eventstore-base.ts
@@ -1,0 +1,6 @@
+export {
+  EventstoreResourceAlreadyExistError,
+  EventstoreResourceNotExistError,
+} from 'src/__mocks__/@reimagined/eventstore-base'
+
+export default jest.fn()